+++ Project.toml
@@ -13,7 +13,7 @@ Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 [compat]
 CommonMark = "0.6"
 Franklin = "0.10"
-FranklinUtils = "0.2"
+FranklinUtils = "0.3"
 NodeJS = "1.1"
 julia = "1.3"
