--- page/Project.toml
@@ -1,4 +1,0 @@
-[deps]
-DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Franklin = "713c75ef-9fc9-4b05-94a9-213340da978e"
-PkgPage = "e7214860-93a8-4f22-b43d-bd447d1a2094"
