+++ page/config.md
@@ -147,4 +147,8 @@ commands that are defined and used in PkgPage.jl
 @def section_counter = 1
 @def showall         = true
 
-\newcommand{\center}[1]{~~~<div style="text-align:center;">~~~#1~~~</div>~~~}
+\newcommand{\html}[1]{~~~#1~~~}
+
+\newenvironment{center}{\html{<div style="text-align:center;">}}{\html{</div>}}
+
+\newenvironment{columns}{\html{<div class="container"><div class="row">}}{\html{</div></div>}}
