+++ src/PkgPage.jl
@@ -9,8 +9,9 @@ using FranklinUtils
 export newpage
 export serve
 
-export lx_lead, lx_figure, lx_table, lx_alert
-export lx_begin, lx_end, lx_column
+export lx_lead, lx_figure, lx_alert
+export env_column, env_table, env_section
+# export lx_begin, lx_end, lx_column
 
 export hfun_col
 
@@ -18,7 +19,8 @@ const F = Franklin
 
 include("instantiate.jl")
 
-include("coms/basic.jl")
-include("coms/begin-end.jl")
+include("latex/basic.jl")
+# include("coms/begin-end.jl")
+include("latex/environments.jl")
 
 end # module
